requirements.txt
Django==4.2.3
googletrans==3.1.0a0
googletrans==4.0.0rc1
newspaper3k==0.2.8
PyPDF2==3.0.1
textblob==0.18.0.post0
transformers==4.47.1
